optmize donotification structure in conversation.
