+++ internal/conversation_msg/delete.go
@@ -27,6 +27,8 @@ import (
 
 	"github.com/jinzhu/copier"
 	pbMsg "github.com/openimsdk/protocol/msg"
+	"github.com/openimsdk/protocol/sdkws"
+	"github.com/openimsdk/tools/errs"
 	"github.com/openimsdk/tools/log"
 )
 
@@ -206,5 +208,45 @@ func (c *Conversation) deleteMessageFromLocal(ctx context.Context, conversationI
 		c.doUpdateConversation(common.Cmd2Value{Value: common.UpdateConNode{Action: constant.ConChange, Args: []string{conversationID}}})
 	}
 	c.msgListener().OnMsgDeleted(utils.StructToJsonString(s))
+	return nil
+}
+
+func (c *Conversation) doDeleteMsgs(ctx context.Context, msg *sdkws.MsgData) error {
+	tips := sdkws.DeleteMsgsTips{}
+	utils.UnmarshalNotificationElem(msg.Content, &tips)
+	log.ZDebug(ctx, "doDeleteMsgs", "seqs", tips.Seqs)
+	for _, v := range tips.Seqs {
+		msg, err := c.db.GetMessageBySeq(ctx, tips.ConversationID, v)
+		if err != nil {
+			log.ZError(ctx, "GetMessageBySeq err", err, "conversationID", tips.ConversationID, "seq", v)
+			continue
+		}
+		var s sdk_struct.MsgStruct
+		copier.Copy(&s, msg)
+		err = c.msgConvert(&s)
+		if err != nil {
+			log.ZError(ctx, "parsing data error", err, "msg", msg)
+			return errs.Wrap(err)
+		}
+		if err := c.deleteMessageFromLocal(ctx, tips.ConversationID, msg.ClientMsgID); err != nil {
+			log.ZError(ctx, "deleteMessageFromLocal err", err, "conversationID", tips.ConversationID, "seq", v)
+			return errs.Wrap(err)
+		}
+	}
+	return nil
+}
+
+func (c *Conversation) doClearConversations(ctx context.Context, msg *sdkws.MsgData) error {
+	tips := &sdkws.ClearConversationTips{}
+	utils.UnmarshalNotificationElem(msg.Content, tips)
+	log.ZDebug(ctx, "doClearConversations", "tips", tips)
+	for _, v := range tips.ConversationIDs {
+		if err := c.clearConversationAndDeleteAllMsg(ctx, v, false, c.db.ClearConversation); err != nil {
+			log.ZError(ctx, "clearConversation err", err, "conversationID", v)
+			return errs.Wrap(err)
+		}
+	}
+	c.doUpdateConversation(common.Cmd2Value{Value: common.UpdateConNode{Action: constant.ConChange, Args: tips.ConversationIDs}})
+	c.doUpdateConversation(common.Cmd2Value{Value: common.UpdateConNode{Action: constant.TotalUnreadMessageChanged}})
 	return nil
 }
